requirements.txt
anthropic>=0.25.0
langgraph>=0.1.0
langchain>=0.2.0
langchain-anthropic>=0.1.0
streamlit>=1.35.0
playwright>=1.44.0
beautifulsoup4>=4.12.0
dnspython>=2.6.0
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0